Migrate to new bring library (#17)
* Migrate bring library and refactor ingredient parsing

* Bump python version

* Bump pre-commit hooks

* Moved the `parse_ignored_ingredients` function from `BringHandler` to `main.py`

* Remove UUID field from Ingredient class initialization

The UUID field is now generated dynamically in the `to_dict` method instead of being set during initialization.

* Fix and refactor ignored ingredient parsing

* Fix erroneous space in specification if unit and quantity are None

* Add support for default values in env variable getter

Updated the `EnvironmentVariableGetter.get` method to accept an optional `default_value` parameter. If the specified environment variable is not set, the method now returns the provided default value instead of raising an error.

* Add support for handling ingredients without amounts.

Introduced `IngredientWithAmountsDisabled` to process ingredients when amounts are not enabled. Updated the ingredient parsing logic to conditionally handle ingredients based on the `enable_amount` flag.

* Remove erroneous space in specification

* Add fix for unparsed recipes

* Removed unnecessary brackets around log messages

* Update README to note Mealie v2+ requirement

Added an important notice in the README to clarify that the integration only supports Mealie version 2 or higher, released in October 2024. Deprecated support for earlier versions is also mentioned with a reference to the relevant pull request.